Feature/exceptions

+++ src/main/java/com/restaurant/api/service/DishService.java
@@ -4,9 +4,12 @@ import com.restaurant.api.dto.request.DishRequest;
 import com.restaurant.api.dto.response.DishCreatedResponse;
 import com.restaurant.api.dto.response.DishListItemResponse;
 import com.restaurant.api.entities.Dish;
+import com.restaurant.api.exception.BadRequestException;
 import com.restaurant.api.exception.ItemNotFoundException;
+import com.restaurant.api.exception.ResourceInUseException;
 import com.restaurant.api.mapper.DishMapper;
 import com.restaurant.api.repository.DishRepository;
+import com.restaurant.api.repository.OrderItemRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +19,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DishService {
     private final DishRepository dishRepository;
+    private final OrderItemRepository orderItemRepository;
     private final DishMapper dishMapper;
 
     public List<DishListItemResponse> getAllDishes() {
@@ -31,6 +35,8 @@ public class DishService {
     }
 
     public DishCreatedResponse addDish(DishRequest dishRequest) {
+        if (dishRepository.existsByName(dishRequest.getName()))
+            throw new BadRequestException("Dish with name: " + dishRequest.getName() + " already exists");
         Dish dish = dishMapper.toEntity(dishRequest);
         Dish saved = dishRepository.save(dish);
         return dishMapper.toCreatedResponse(saved);
@@ -39,6 +45,11 @@ public class DishService {
     public DishCreatedResponse updateDish(Long id, DishRequest dishRequest) {
         Dish dish = dishRepository.findById(id)
                 .orElseThrow(() -> new ItemNotFoundException(Dish.class, id));
+
+        // Check if new name is not taken by another dish
+        if (!dish.getName().equals(dishRequest.getName()) && dishRepository.existsByName(dishRequest.getName()))
+            throw new BadRequestException("Dish with name: " + dishRequest.getName() + " already exists");
+
         dishMapper.updateFromRequest(dish, dishRequest);
         Dish saved = dishRepository.save(dish);
         return dishMapper.toCreatedResponse(saved);
@@ -47,6 +58,10 @@ public class DishService {
     public void deleteDish(Long id) {
         Dish dish = dishRepository.findById(id)
                 .orElseThrow(() -> new ItemNotFoundException(Dish.class, id));
+
+        if(orderItemRepository.existsByDishId(id))
+            throw new ResourceInUseException("Cannot delete dish because it is part of existing orders");
+
         dishRepository.delete(dish);
     }
 }
